### 2023-03-06 #5219 @kyleniemeyer
Hi @hgrecco, thanks for resubmitting with the paper in the software repo, I was going to ask you to do just that!

It does look like there are some issues with the paper, specifically in the YAML header associated with one of the author's affiliations. Can you fix those tabs to be spaces like the rest of the YAML?
