### 2023-03-23 #5219 @lucask07
Hi @matthewfeickert, sorry I need to decline this offer due to limited time availability. PyVISA is great and I use it all the time! Thanks to the PyVISA team.
