### 2023-03-26 404ef1f
Use architecture instead of `bitness` in `LibraryPath` and friends
Fixes #655

- Replaced `bitness` of libraries and python interpreter
- Removed unused `LibraryPath.is32bit`, `LibraryPath.is_64bit`, and
  `LibraryPath.bitness` methods
- Introduced AARCH64 as tested-for architecture
- Slightly updated detection logic
- Removed IA64 and linux universal binary test scenarios since they are
  unrealistic
- Removed `linux2/3` checking since it should not exist past Python 3.3
