+++ pyvisa/util.py
@@ -18,6 +18,8 @@ import subprocess
 import sys
 import warnings
 from collections import OrderedDict
+from enum import Enum
+from pathlib import Path
 from types import ModuleType
 from typing import (
     Any,
@@ -161,6 +163,56 @@ def add_user_dll_extra_paths() -> Optional[List[str]]:
         return None
 
 
+class PEMachineType(Enum):
+    UNKNOWN = 0
+    I386 = 0x014C
+    R3000 = 0x0162
+    R4000 = 0x0166
+    R10000 = 0x0168
+    WCEMIPSV2 = 0x0169
+    ALPHA = 0x0184
+    SH3 = 0x01A2
+    SH3DSP = 0x01A3
+    SH3E = 0x01A4
+    SH4 = 0x01A6
+    SH5 = 0x01A8
+    ARM = 0x01C0
+    AARCH64 = 0xAA64
+    THUMB = 0x01C2
+    ARMNT = 0x01C4
+    AM33 = 0x01D3
+    POWERPC = 0x01F0
+    POWERPCFP = 0x01F1
+    IA64 = 0x0200
+    MIPS16 = 0x0266
+    ALPHA64 = 0x0284
+    MIPSFPU = 0x0366
+    MIPSFPU16 = 0x0466
+    TRICORE = 0x0520
+    CEF = 0x0CEF
+    EBC = 0x0EBC
+    AMD64 = 0x8664
+    M32R = 0x9041
+    CEE = 0xC0EE
+
+
+class ArchitectureType(Enum):
+    I386 = ("x86", 32)
+    X86_64 = ("x86", 64)
+    AARCH64 = ("arm", 64)
+
+    @classmethod
+    def from_platform_machine(cls, machine: str) -> Optional["ArchitectureType"]:
+        if machine == "i386" or machine == "i686":
+            return cls.I386
+        elif machine == "x86_64" or machine == "amd64":
+            return cls.X86_64
+        elif machine == "arm64" or machine == "aarch64":
+            return cls.AARCH64
+
+        return None
+
+
 class LibraryPath(str):
     """Object encapsulating information about a VISA dynamic library."""
 
@@ -171,7 +223,7 @@ class LibraryPath(str):
     found_by: str
 
     #: Architectural information (32, ) or (64, ) or (32, 64)
-    _arch: Optional[Tuple[int, ...]] = None
+    _arch: Optional[list[ArchitectureType]] = None
 
     def __new__(
         cls: Type["LibraryPath"], path: str, found_by: str = "auto"
@@ -183,36 +235,15 @@ class LibraryPath(str):
         return obj
 
     @property
-    def arch(self) -> Tuple[int, ...]:
+    def arch(self) -> list[ArchitectureType]:
         """Architecture of the library."""
         if self._arch is None:
             try:
-                self._arch = get_arch(self.path)
+                self._arch = get_arch(Path(self.path))
             except Exception:
-                self._arch = tuple()
+                self._arch = []
 
         return self._arch
-
-    @property
-    def is_32bit(self) -> Union[bool, Literal["n/a"]]:
-        """Is the library 32 bits."""
-        if not self.arch:
-            return "n/a"
-        return 32 in self.arch
-
-    @property
-    def is_64bit(self) -> Union[bool, Literal["n/a"]]:
-        """Is the library 64 bits."""
-        if not self.arch:
-            return "n/a"
-        return 64 in self.arch
-
-    @property
-    def bitness(self) -> str:
-        """Bitness of the library."""
-        if not self.arch:
-            return "n/a"
-        return ", ".join(str(a) for a in self.arch)
 
 
 def cleanup_timeout(timeout: Optional[Union[int, float]]) -> int:
[…]
