### 2023-03-26 2bd18b7
Fix dataclass hierarchy uncovered by mypy
