### 2023-03-26 aebdfe5
Fix dataclass hierarchy uncovered by mypy (#734)
* Fix dataclass hierarchy uncovered by mypy

* Remove __future__.annotations and other unrelated changes

* Revert missed change
