+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@ repos:
     hooks:
       - id: isort
   - repo: https://github.com/psf/black
-    rev: 22.10.0
+    rev: 23.1.0
     hooks:
       - id: black
   - repo: https://github.com/pycqa/flake8
@@ -12,7 +12,7 @@ repos:
     hooks:
       - id: flake8
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.991 # Use the sha / tag you want to point at
+    rev: v1.1.1 # Use the sha / tag you want to point at
     hooks:
       - id: mypy
         additional_dependencies: [numpy, typing_extensions]
