### 2023-03-26 578a903
Fix typehints to make them py3.8-compatible
