### 2023-03-27 #5304 @editorialbot
[REVIEW]: PyVISA: the Python instrumentation package
**Submitting author:** @hgrecco (<a href="http://orcid.org/0000-0002-1165-4320">Hernan Edgardo Grecco</a>)
**Repository:** https://github.com/pyvisa/pyvisa
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.13.0
