### 2023-03-27 2ded384
Update Path-annotated arguments to be less restrictive
