### 2023-03-31 #5304 @aquilesC
As a very frequent user of PyVISA, I’m glad to see it becoming a paper. The authors have done a great job over many years.

After reviewing the code and paper, I conclude it is ready for publication.

I only found a typo on the manuscript on line 31 it should read receive instead of receives.
