### 2023-03-31 #5304 @sidihamady
Even if at the beginning PyVISA was created as a light wrapper for the VISA shared library C functions, its wide adoption, its simplicity of use and the combination with the many scientific packages of Python made it a must. I use it both for teaching and to prototype characterization tools. Its code is particularly well organized, well documented and the examples are fully functional. It works very well with VISA backends from NI and Keysight. It's a bit harder to use non-proprietary backends, but it's feasible. Authors indicate that will include async IO in a next release, which is particularly interesting for instrumentation but challenging to implement in Python and to use. Another task could be to develop a simulator more usable and less limited than PyVISA-sim, since it is difficult to quickly develop and debug without being continuously connected to real instruments.
I recommend its publication in JOSS.
