### 2023-04-14 6d2791e
resources: Update context manager type annotation to preserve derived type (#740)
* resources: Update context manager type annotation to preserve derived type

The type annotation for Resource.__enter__() indicates that it returns the base
class, Resource. If you narrow the resource type to a derived class before
entering the context manager, entering the context manager casts it back to the
base class.

The solution is to use a type var to specify that the return type is the same as
the type of `self`.

* Update changelog
