### 2023-04-21 #5304 @hgrecco
There is a version published in a journal. Sh

Raymond, E. The cathedral and the bazaar. Know Techn Pol 12, 23–49 (1999). https://doi.org/10.1007/s12130-999-1026-0

but the official site is this:

http://www.catb.org/~esr/writings/cathedral-bazaar/cathedral-bazaar/

I leave it to the editor to decide what is the most suitable.
