### 2023-04-22 #5304 @matthewfeickert
> MISSING DOIs
>
> - 10.4135/9781412950657.n33 may be a valid DOI for title: The Cathedral and the Bazaar

This is the DOI for a [chapter by the same title](https://sk.sagepub.com/reference/newmedia/n33.xml) by Theresa M. Senft in the book [Encyclopedia of New Media](https://sk.sagepub.com/reference/newmedia), and so is fine to ignore. :+1:
