### 2023-04-22 #5304 @matthewfeickert
> I leave it to the editor to decide what is the most suitable.

Ah, sorry missed this before posting given some lag in cached pages. :/ Having _some_ DOI associated is perferable, but the rest of the citation is fine. Let me PR this quickly.
