### 2023-04-24 ffc63df
chore: Update 'Cathedral and the Bazaar' reference
* Use the published Springer article as it has a DOI and is archive
  stable.
   - Note that the text probably referenced was from the authors website
     http://www.catb.org/~esr/writings/cathedral-bazaar/
