+++ paper.bib
@@ -55,20 +55,18 @@
   eprint = {https://www.science.org/doi/pdf/10.1126/science.215.4534.785},
 }
 
-@book{raymond2010cathedral,
-  added-at = {2016-06-06T09:35:55.000+0200},
-  author = {Raymond, Eric Steven},
-  biburl = {https://www.bibsonomy.org/bibtex/200347d9734fc1da1e457bc3fd5df344c/dirtyhawk},
-  interhash = {ef5b3f6887f9cedc384017578644f8fc},
-  intrahash = {00347d9734fc1da1e457bc3fd5df344c},
-  isbn = {978-1607962281},
-  keywords = {open_source programming},
-  publisher = {Snowball Publishing},
-  timestamp = {2016-06-06T09:36:12.000+0200},
-  title = {The Cathedral and the Bazaar},
-  url = {http://www.catb.org/~esr/writings/cathedral-bazaar/},
-  urldate = {2016-06-06},
-  year = 2010
+@article{raymond:1999,
+	title = {The Cathedral and the Bazaar},
+	volume = {12},
+	issn = {1874-6314},
+	url = {https://doi.org/10.1007/s12130-999-1026-0},
+	doi = {10.1007/s12130-999-1026-0},
+	language = {en},
+	number = {3},
+	urldate = {2023-04-24},
+	journal = {Know Techn Pol},
+	author = {Raymond, Eric},
+	month = sep,
+	year = {1999},
+	pages = {23--49},
 }
-
-
