### 2023-04-25 #5304 @matthewfeickert
@hgrecco all that's left to do now before publication is to ensure that there is a long term public archive of the code that was reviewed. In this case the code was [`v1.13.0`](https://github.com/pyvisa/pyvisa/releases/tag/1.13.0) though there has been [some development between then and the current `HEAD`](https://github.com/pyvisa/pyvisa/compare/1.13.0...6d2791e195c169cb680863ddcae487cc17364eee).

We'd suggest depositing the code either with [Zenodo](https://zenodo.org/) or with [figshare](https://figshare.com/) to get an archive with a DOI. If you use Zenodo there is [a GitHub integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) that can create a Zenodo archive for you anytime you make a GitHub release of your code.

[![Zenodo_DOI_guide](https://user-images.githubusercontent.com/5142394/234173148-2226325d-0a42-4526-b546-75b469d26c4c.png)](https://zenodo.org/account/settings/github/)

If you'd like to make a new release of PyVISA to trigger the uploader that's fine and we can have @editorialbot update the version listed in the review. You're also welcome to not do that and just upload the state of the repository at [`v1.13.0`](https://github.com/pyvisa/pyvisa/releases/tag/1.13.0). Either way we'll just need you to share the resulting DOI of the archive created in this thread.

Let me know if you have any questions. :+1:
