### 2023-04-26 8fda73d
mention session sharing in ResourceManager.close docstring
