+++ pyvisa/highlevel.py
@@ -3063,7 +3063,15 @@ class ResourceManager(object):
         return self.visalib.get_last_status_in_session(self.session)
 
     def close(self) -> None:
-        """Close the resource manager session."""
+        """Close the resource manager session.
+
+        Notes
+        -----
+        Since the resource manager session is shared between instances
+        this will also terminate connections obtained from other
+        ResourceManager instances.
+
+        """
         atexit.unregister(self._atexit_handler)
         try:
             logger.debug("Closing ResourceManager (session: %s)", self.session)
